sensorlogger/requirements.txt
aniso8601==9.0.1
click==8.1.3
colorama==0.4.5
Flask==2.2.2
Flask-RESTful==0.3.9
itsdangerous==2.1.2
Jinja2==3.1.2
mariadb==1.0.11
MarkupSafe==2.1.1
PyJWT==2.5.0
pytz==2022.4
six==1.16.0
Werkzeug==2.2.2
